[taxi] added possibility to use mwm ids as enabled/disabled territories
