+++ partners_api/CMakeLists.txt
@@ -29,11 +29,11 @@ set(
   rb_ads.hpp
   taxi_base.cpp
   taxi_base.hpp
-  taxi_countries.cpp
-  taxi_countries.hpp
+  taxi_places.cpp
+  taxi_places.hpp
   taxi_engine.cpp
   taxi_engine.hpp
-  taxi_places.hpp 
+  taxi_places_constants.hpp
   taxi_provider.hpp
   uber_api.cpp
   uber_api.hpp
