+++ partners_api/partners_api_tests/taxi_engine_tests.cpp
@@ -25,6 +25,7 @@ public:
   storage::TCountriesVec GetCountryIds(ms::LatLon const & latlon) override { return {"Belarus"}; }
 
   std::string GetCityName(ms::LatLon const & latlon) override { return "Minsk"; }
+  storage::TCountryId GetMwmId(ms::LatLon const & latlon) override { return ""; }
 };
 
 class UkraineOdessaDelegate : public taxi::Delegate
@@ -33,6 +34,7 @@ public:
   storage::TCountriesVec GetCountryIds(ms::LatLon const & latlon) override { return {"Ukraine"}; }
 
   std::string GetCityName(ms::LatLon const & latlon) override { return "Odessa"; }
+  storage::TCountryId GetMwmId(ms::LatLon const & latlon) override { return ""; }
 };
 
 class UkraineMariupolDelegate : public taxi::Delegate
@@ -41,6 +43,7 @@ public:
   storage::TCountriesVec GetCountryIds(ms::LatLon const & latlon) override { return {"Ukraine"}; }
 
   std::string GetCityName(ms::LatLon const & latlon) override { return "Mariupol"; }
+  storage::TCountryId GetMwmId(ms::LatLon const & latlon) override { return ""; }
 };
 
 class BulgariaSofiaDelegate : public taxi::Delegate
@@ -49,6 +52,7 @@ public:
   storage::TCountriesVec GetCountryIds(ms::LatLon const & latlon) override { return {"Bulgaria"}; }
 
   std::string GetCityName(ms::LatLon const & latlon) override { return "Sofia"; }
+  storage::TCountryId GetMwmId(ms::LatLon const & latlon) override { return ""; }
 };
 
 class UsaDelegate : public taxi::Delegate
@@ -60,6 +64,7 @@ public:
   }
 
   std::string GetCityName(ms::LatLon const & latlon) override { return ""; }
+  storage::TCountryId GetMwmId(ms::LatLon const & latlon) override { return ""; }
 };
 
 class RussiaKonetsDelegate : public taxi::Delegate
@@ -71,6 +76,7 @@ public:
   }
 
   std::string GetCityName(ms::LatLon const & latlon) override { return "Konets"; }
+  storage::TCountryId GetMwmId(ms::LatLon const & latlon) override { return ""; }
 };
 
 std::vector<taxi::Product> GetUberSynchronous(ms::LatLon const & from, ms::LatLon const & to,
