+++ partners_api/taxi_base.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "partners_api/taxi_countries.hpp"
+#include "partners_api/taxi_places.hpp"
 #include "partners_api/taxi_provider.hpp"
 
 #include <functional>
@@ -53,11 +53,11 @@ using ApiPtr = std::unique_ptr<ApiBase>;
 
 struct ApiItem
 {
-  ApiItem(Provider::Type type, ApiPtr && api, Countries const & enabled, Countries const & disabled)
+  ApiItem(Provider::Type type, ApiPtr && api, Places const & enabled, Places const & disabled)
     : m_type(type)
     , m_api(std::move(api))
-    , m_enabledCountries(enabled)
-    , m_disabledCountries(disabled)
+    , m_enabledPlaces(enabled)
+    , m_disabledPlaces(disabled)
   {
   }
 
@@ -65,11 +65,13 @@ struct ApiItem
                                std::string const & city) const;
   bool IsAnyCountryEnabled(storage::TCountriesVec const & countryIds,
                            std::string const & city) const;
+  bool IsMwmDisabled(storage::TCountryId const & mwmId) const;
+  bool IsMwmEnabled(storage::TCountryId const & mwmId) const;
 
   Provider::Type m_type;
   ApiPtr m_api;
 
-  Countries m_enabledCountries;
-  Countries m_disabledCountries;
+  Places m_enabledPlaces;
+  Places m_disabledPlaces;
 };
 }  // namespace taxi
